* Add new API for adding, removing extra multicast groups, together with IGMPv3 filtering. * Update examples with latest pgm_transport_create parameters.

+++ trunk/openpgm/pgm/examples/pgmsend.c
@@ -150,7 +150,7 @@ create_transport (void)
 	g_assert (e == 0);
 
 	struct group_source_req recv_gsr, send_gsr;
-	int recv_len = 1;
+	gsize recv_len = 1;
 	e = pgm_if_parse_transport (g_network, AF_UNSPEC, &recv_gsr, &recv_len, &send_gsr);
 	g_assert (e == 0);
 	g_assert (recv_len == 1);
